Add files via upload
A hypersonic weapon is a weapon capable of travelling at hypersonic speed, defined as between 5 and 25 times the speed of sound or about 1 to 5 miles per second. Hypersonic glide vehicle (HGV) is a missile warheads which maneuver and glide through the atmosphere at high speeds after an initial ballistic launch phase. In this python project notebook we will analize attack power with machine learning algorithm